requirements.txt
autopep8==1.6.0
black==21.12b0
click==8.0.3
cycler==0.11.0
fonttools==4.28.3
kiwisolver==1.3.2
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
pep8==1.7.1
Pillow==8.4.0
pkg_resources==0.0.0
platformdirs==2.4.0
pycodestyle==2.8.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.3
six==1.16.0
toml==0.10.2
tomli==1.2.3
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
typing_extensions==4.0.1
